vision-py/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "battle-vision-overlay-vision"
version = "0.1.0"
description = "Bootstrap package for battle-vision-overlay vision module"
requires-python = ">=3.11"
dependencies = ["Pillow>=12,<13"]

[project.scripts]
vision-crop-poc = "vision.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"vision.ocr" = ["*.swift"]
